+++ src/types/env.d.ts
@@ -9,6 +9,6 @@ interface ImportMeta {
 }
 
 interface Window {
-  dataLayer: unknown[];
+  dataLayer: Array<Record<string, unknown> | unknown[]>;
   gtag: (...args: unknown[]) => void;
 }